refactor: 개별 차량 경로 조회 API에서 위경도 데이터타입 수정

+++ monicar-control-center/src/main/java/org/controlcenter/vehicle/presentation/dto/RouteResponse.java
@@ -1,11 +1,10 @@
 package org.controlcenter.vehicle.presentation.dto;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record RouteResponse(
-	BigDecimal lat,
-	BigDecimal lng,
+	Integer lat,
+	Integer lng,
 	Integer spd,
 	LocalDateTime timestamp
 ) {
